move behaviourscript to seperate folder
